Se arregla un detalle del ejercicio1

+++ tp1_AndreuLuca/src/ar/edu/unju/fi/ejercicio1/calculadora.java
@@ -2,8 +2,8 @@ package ar.edu.unju.fi.ejercicio1;
 public class calculadora {
 	public calculadora() {
 	}
-	public int SumarYpromediar(int val1, int val2, int val3, int val4, int val5) {
-		int resultado=val1+val2+val3+val4+val5;
+	public float SumarYpromediar(int val1, int val2, int val3, int val4, int val5) {
+		float resultado=val1+val2+val3+val4+val5;
 		return resultado/5;
 	}
 }
